Linux's temperature support has been added.
